+++ HISTORY.rst
@@ -5,4 +5,12 @@ History
 0.1.0 (2020-05-27)
 ------------------
 
-* First release on PyPI.
+* First release.
+
+0.1.1 (2020-05-27)
+------------------
+
+* Black formatting
+* Remove FASTQ input file option
+* Default to write to a SAM file instead of a BAM file.
+* Documentation